Use latest method on iOS 16+; refactor tests
